build system for perf/C and perf/C++

+++ perf/cpp/local_lat.cpp
@@ -22,6 +22,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <assert.h>
+#include <stddef.h>
 
 int main (int argc, char *argv [])
 {
@@ -32,7 +33,7 @@ int main (int argc, char *argv [])
     }
     const char *bind_to = argv [1];
     int roundtrip_count = atoi (argv [2]);
-    int message_size = atoi (argv [3]);
+    size_t message_size = (size_t) atoi (argv [3]);
 
     zmq::context_t ctx (1, 1);
